pyproject.toml
[project]
name = "doc-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "certifi>=2026.2.25",
    "fastapi>=0.135.3",
    "firecrawl-py>=4.22.1",
    "isort>=8.0.1",
    "jinja2>=3.1.6",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchainhub>=0.1.21",
    "langsmith>=0.7.26",
    "nltk>=3.9.4",
    "openai>=2.30.0",
    "pillow>=12.2.0",
    "protobuf==3.20.*",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "streamlit-chat>=0.1.1",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "unstructured>=0.22.16",
    "uvicorn>=0.44.0",
]
